requirements.txt
PyPDF2==3.0.1
python-docx==1.1.0
colorama==0.4.6